finished rewriting of handler system
CHANGES
=======

The work of assigining a handler is now in dhandlers.h. The function
init_handlers() sets up a linked list of Matcher objects. These are
tuples of a match function and a handler function. Assign_handlers
goes through the list of matchers running the match function for each
one. It stops when reaching a matcher that returns true. At this point,
the matcher should have also set the handler for the request.

This still looks weird. I'll try to simplify it.