restyling: removed #ifndef dance
CHANGES
=======

A way to solve the problem of redefining functions by including the same
header several times is to enclose the said header in a

| #ifndef HEADER_H
| #define HEADER_H
| ...
| #endif

Now this isn't particularly bad. But there's an easier way to do it,
namely just include the files in the correct order. That's what this
commit does.

Dreadful idea, isn't it?