added experimental resource system
The idea, so far, is this: you have some textfiles that you'd like
included in the source; but they require some preprocessing (such as
escaping "s and so on), or it's just hard to use the cpp; what do you
do? You use drc (Dissemina Resource Compiler).

Drc is a simple Perl script. It takes one argument -- the file to be
processed. It reads this file, line by line, replacing all occurrences
of the pattern @@<filename>@@ with <content-of-file>. This allows for
easy inclusion of, say, html page templates.

Note that, so far, dissemina hasn't been modified to make use of this
system.