finished switch to new network IO system
Before, TCP multiplexing was done by the individual handlers. This was
bad. For example, some handlers, such as the directory lister and the
error handler, didn't bother to do it at all. This looked ugly and it
made no sense. Another problem was that none of the handlers did any
multiplexing when sending out headers.

The new system is slightly better at this: handlers use a few well-defined
functions to create ``envelopes''. Dissemina then cycles through a list
of envelopes and sends out some data for each. This allows the handlers
to worry about their on problems and not about network IO.

It's also worth noting that the said functions (which reside in dstdio.h)
use the blocking version of the send() functions, so it's unlikely that
any multiplexing will be done.

This commit is just include a working version of this system in the
history. There might be bugs and there might be memory leaks, but all
looks well... for now.