cleaned up Makefile
Make executes the rules for a target sequentially and stops at the first
error. But what is an error. Apparently, any command-line that returns
non-zero is considered an error. This is not always the desired behaviour.

Take, for example, the pkill command. Given a name, it finds a process
that matches and kills it. And if it doesn't? It returns non zero.

I would short-circuit this problem by writing the command as ``pkill
dissemina || true'', but this was ugly. The new lines in Makefile tell
make to ignore errors on the specified targets.