bugfixes and new request parsing system
Perhaps ``parsing system'' is a bit of an overstatement. The parsing of
requests is now done by the parsereq function. It's the direct successor
of fill_in_request. Now it resides in drequest.h and has been cleaned
up. (ie no more leftovers from who knows when)

I've also marked Bug 4 as solved. I forgot to do this last release.