moved from MIT to GPL3
Free as in Freedom, Richard Stallman's biography, won me over.

Dissemina is free, but it's not public domain. You can download it and
you can redistribute it. But the price you pay is that any modification
you make to the source you must make free.

Update LICENSE file. Updated file headers and README.