added preliminary support for CGI
CGI can be summed up as: set some environment variables, run a command
and send back it's output. Dissemina can now do the second step. By
requesting a URL of the form /call/foo, dissemina executes foo and
returns its output as plain text. It's not elegant, but it'll do for now.